Modification production.c pour le calcul du rendement geographique pour les valeurs qui ne sont pas dans le tableau + tests associés

+++ trunk/ projet-ia/inc/structures.h
@@ -152,7 +152,7 @@ typedef struct donne_geo
 {
   int inclinaison; //Inclinaison du toit en degré
   int orientation; //Orientation de la maison °/sud
-  int rendement; //Redement géographique
+  float rendement; //Redement géographique
 }ST_DonneGeo;
 
 #endif
